Completed requested changes in teams
